feat: 충돌 해결 및 pendingBooking 관련 로직을 pendingBookingService로 이관 (#57)
